pyproject.toml
[project]
name = "llm-experiment-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.45.0",
]
